pyproject.toml
[project]
name = "cross-auth"
version = "0.1.0"
description = "A modern Authentication Framework for Python"
readme = "README.md"
authors = [{ name = "Patrick Arminio", email = "patrick.arminio@gmail.com" }]
requires-python = ">=3.11"
dependencies = [
 "lia-web>=0.1.2",
 "passlib[bcrypt]>=1.7.4",
 "pydantic[email]>=2.11.3",
 "bcrypt>=4.0.0,<5.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "fastapi>=0.115.12",
    "httpx>=0.28.1",
    "pytest>=7.4.4",
    "pytest-asyncio>=0.23.8",
    "python-multipart>=0.0.20",
    "starlette>=0.46.1",
    "inline-snapshot>=0.20.10",
    "respx>=0.22.0",
    "coverage>=7.8.0",
    "time-machine>=2.16.0",
]


[tool.inline-snapshot]
format-command = "ruff format --stdin-filename {filename}"
